Seperate tests into their respective suites
